+++ src/AtelierPlugin.php
@@ -7,6 +7,7 @@ namespace Safi\Atelier;
 use Filament\Contracts\Plugin;
 use Filament\Panel;
 use Safi\Atelier\Filament\Pages\PageEditor;
+use Safi\Atelier\Filament\Pages\SiteDetails;
 use Safi\Atelier\Filament\Resources\PageResource;
 
 class AtelierPlugin implements Plugin
@@ -62,6 +63,7 @@ class AtelierPlugin implements Plugin
             ])
             ->pages([
                 PageEditor::class,
+                SiteDetails::class,
             ]);
     }