+++ src/AtelierServiceProvider.php
@@ -22,6 +22,7 @@ class AtelierServiceProvider extends PackageServiceProvider
                 'create_atelier_tables',
                 'create_atelier_page_revisions_table',
                 'create_atelier_page_redirects_table',
+                'create_atelier_settings_table',
             ]);
     }